commenting controllers

+++ app/Http/Controllers/CartController.php
@@ -11,26 +11,47 @@ use Illuminate\Http\JsonResponse;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 
+/**
+ * This class handles all the cart related functions
+ */
 class CartController extends Controller
 {
 
-    public function getAllCarts(Request $request)
+    /** This function gets all the items in the cart of the logged in user
+     *  and displays them in the cart page
+     *
+     * @return Factory|View|RedirectResponse|Application
+     */
+    public function getAllCarts(): Factory|View|RedirectResponse|Application
     {
+        // get all the items in the cart of the logged-in user
+        $carts = Cart::where('client_id', auth()->user()->id)->get();
 
-        $client_id = auth()->user()->id;
-
-        $carts = Cart::where('client_id', $client_id)->get();
+        // put the number of items in the cart in the session
         session()->put('cartCount', $carts->count());
 
+        // if the cart is empty, redirect to the shop page with an error message
         if ($carts->count() == 0) {
             return redirect()->route("shop")->with('error', 'Your cart is empty');
         }
 
+        // return the cart view with the cart items
         return view('cart', compact('carts'));
     }
 
+    /**
+     * This function adds a product to the cart
+     * this function is called via ajax
+     * and returns a json response
+     *
+     *
+     * @param Request $request
+     * @return JsonResponse
+     */
     public function addToCart(Request $request): JsonResponse
     {
+        // making use of the laravel validation
+        // validate the request for the required fields
         $request->validate([
             'client_id' => 'required',
             'product_id' => 'required',
@@ -38,7 +59,19 @@ class CartController extends Controller
             'size' => 'required',
         ]);
 
-        $exist = Cart::where('client_id', $request->client_id)->where('product_id', $request->product_id)->where('size', $request->size)->first();
+        $clientId = $request->client_id;
+
+        // validate the client_id is the same as the logged-in user
+        // this is to prevent a user from adding a product to another user's cart
+        if ($clientId != auth()->user()->id) {
+            return response()->json([
+                'message' => 'You are not authorized to add this product to cart',
+            ], 401);
+        }
+
+        // check if the product already exists in the cart
+        // if it exists, update the quantity
+        $exist = Cart::where('client_id', $clientId)->where('product_id', $request->product_id)->where('size', $request->size)->first();
 
         if ($exist) {
             $exist->quantity += $request->quantity;
@@ -49,6 +82,7 @@ class CartController extends Controller
             ]);
         }
 
+        // if the product does not exist in the cart, create a new cart item
         $cart = Cart::create([
             'client_id' => $request->client_id,
             'product_id' => $request->product_id,
@@ -56,15 +90,27 @@ class CartController extends Controller
             'size' => $request->size,
         ]);
 
+        // put the number of items in the cart in the session
         $cartCount = Cart::where('client_id', $request->client_id)->count();
         session()->put('cartCount', $cartCount);
 
+        // return a json response
         return response()->json([
             'message' => 'Product added to cart successfully',
             'cart' => $cart,
         ]);
     }
 
+    /**
+     * This function removes a product from the cart
+     * this function is called via ajax
+     * and returns a json response
+     *
+     *
+     * @param Request $request
+     * @param $id
+     * @return JsonResponse
+     */
     public function removeFromCart(Request $request, $id): JsonResponse
     {
         $cart = Cart::find($id);
@@ -75,8 +121,18 @@ class CartController extends Controller
             ], 404);
         }
 
+        // validate the client_id is the same as the logged-in user
+        // this is to prevent a user from removing a product from another user's cart
+        if ($cart->client_id != auth()->user()->id) {
+            return response()->json([
+                'message' => 'You are not authorized to remove this product from cart',
+            ], 401);
+        }
+
+        // delete the cart item
         $cart->delete();
 
+        // put the number of items in the cart in the session
         $cartCount = Cart::where('client_id', $request->client_id)->count();
         session()->put('cartCount', $cartCount);
 
@@ -85,7 +141,12 @@ class CartController extends Controller
         ]);
     }
 
-    public function checkout(Request $request): Factory|View|RedirectResponse|Application
+    /**
+     * This function creates a new order for the user
+     * and redirects to the checkout page
+     * @return Factory|View|RedirectResponse|Application
+     */
+    public function checkout(): Factory|View|RedirectResponse|Application
     {
         $client_id = auth()->user()->id;
 
@@ -97,14 +158,18 @@ class CartController extends Controller
             return redirect()->route('profile.index')->with('error', 'Please update your profile with your address before checkout');
         }
 
+        // get all the items in the cart of the logged-in user
         $carts = Cart::where('client_id', $client_id)->get();
 
+        // calculate the total amount of the order
+        // by multiplying the price of each product by the quantity
         $total_amount = $carts->sum(function ($cart) {
             $cart->product->quantity -= $cart->quantity;
             $cart->product->save();
             return $cart->product->price * $cart->quantity;
         });
 
+        // create a new order for the user
         $order = Order::create([
             'client_id' => $client_id,
             'total_amount' => $total_amount,
@@ -113,6 +178,7 @@ class CartController extends Controller
             'payment_status' => 'pending',
         ]);
 
+        // create a new order item for each cart item
         foreach ($carts as $cart) {
             $order->items()->create([
                 'product_id' => $cart->product_id,
@@ -121,20 +187,42 @@ class CartController extends Controller
             ]);
         }
 
+        // delete all the cart items because
+        // they have been converted to order items
         $carts->each->delete();
+
+        // put the number of items in the cart in the session
         session()->put('cartCount', 0);
 
         return view('checkout', compact('order'));
     }
 
+    /**
+     * This function updates the quantity of a product in the cart
+     * this function is called via ajax
+     * and returns a json response
+     *
+     *
+     * @param Request $request
+     * @return JsonResponse|void
+     */
     public function updateCart(Request $request)
     {
+        // validate the request for the required fields
         $request->validate([
             'cart_id' => 'required',
             'quantity' => 'required|integer|min:1',
         ]);
 
-        $exist = Cart::findOrFail($request->cart_id);
+        $exist = Cart::find($request->cart_id);
+
+        // validate the client_id is the same as the logged-in user
+        // this is to prevent a user from updating a product in another user's cart
+        if ($exist->client_id != auth()->user()->id) {
+            return response()->json([
+                'message' => 'You are not authorized to update this product in cart',
+            ], 401);
+        }
 
         if ($exist) {
             $exist->quantity = $request->quantity;
@@ -143,11 +231,20 @@ class CartController extends Controller
             return response()->json([
                 'message' => "Product quantity updated successfully",
             ]);
+        } else {
+            return response()->json([
+                'message' => "Product not found",
+            ]);
         }
     }
 
 
-    public function confirmation($orderNumber)
+    /**
+     * This function returns the order confirmation page
+     * @param $orderNumber
+     * @return Factory|View|Application
+     */
+    public function confirmation($orderNumber): View|Factory|Application
     {
         $order = Order::findOrFail($orderNumber);
 
@@ -155,8 +252,6 @@ class CartController extends Controller
             'status' => 'confirmed',
         ]);
 
-
         return view('confirmation', compact('order'));
     }
-
 }
